renames files to snake_case
