+++ source/tests/yats-test/scheduler_test.cpp
@@ -1,7 +1,7 @@
 #include <gmock/gmock.h>
 
-#include <yats/Scheduler.h>
-#include <yats/Util.h>
+#include <yats/scheduler.h>
+#include <yats/util.h>
 
 TEST(scheduler_test, simple_create)
 {
