+++ source/yats/include/yats/connection_helper.h
@@ -3,9 +3,9 @@
 #include <map>
 #include <memory>
 
-#include <yats/InputConnector.h>
-#include <yats/OutputConnector.h>
-#include <yats/TaskContainer.h>
+#include <yats/input_connector.h>
+#include <yats/output_connector.h>
+#include <yats/task_container.h>
 
 namespace yats
 {
